WICRPORTAL-56: Add authorisation rule for api requests.
Adds authorisation rule for any URL matching the api (such as /api/cancertypes and /api/clinicalattributes/patients) so that access is granted logged in or requesting from localhost. This enables the dataset validator script to communicate with the portal instance, when the validator script is run on the same machine as the running portal instance.